# introduce a new method (set_host_to_base_host) to replace set_host when set_host is unnecessary